+++ src/app/[locale]/contribute/FeedSubmission/Form/FourthStep.tsx
@@ -14,9 +14,12 @@ import {
   useForm,
   useWatch,
 } from 'react-hook-form';
-import { type YesNoFormInput, type FeedSubmissionFormFormInput } from '.';
 import { useTranslations } from 'next-intl';
 import FormLabelDescription from './components/FormLabelDescription';
+import {
+  type FeedSubmissionFormFormInput,
+  type YesNoFormInput,
+} from './types';
 
 export interface FeedSubmissionFormInputFourthStep {
   dataProducerEmail?: string;
