+++ src/app/[locale]/contribute/FeedSubmission/Form/types.ts
@@ -1,0 +1,1 @@
+export * from '../../../../utils/feed-submission-types';
