+++ src/app/[locale]/contribute/submitted/page.tsx
@@ -1,0 +1,8 @@
+import { type ReactElement } from 'react';
+import FeedSubmitted from '../../../screens/FeedSubmitted';
+
+export default function ContributeSubmittedPage(): ReactElement {
+  return (
+    <FeedSubmitted />
+  );
+}
