+++ src/app/screens/FeedSubmitted.tsx
@@ -1,8 +1,9 @@
+'use client';
+
 import {
   Typography,
   Box,
   Container,
-  CssBaseline,
   useTheme,
   Button,
 } from '@mui/material';
@@ -11,7 +12,6 @@ export default function FeedSubmitted(): React.ReactElement {
   const theme = useTheme();
   return (
     <Container component='main' sx={{ my: 0, mx: 'auto' }} maxWidth='lg'>
-      <CssBaseline />
       <Typography
         data-cy='feedSubmitSuccess'
         component='h1'
