+++ src/app/services/feeds/add-feed-form-service.ts
@@ -1,6 +1,6 @@
 import { getFunctions, httpsCallable } from 'firebase/functions';
 import { app } from '../../../firebase';
-import { type FeedSubmissionFormFormInput } from '../../screens/FeedSubmission/Form';
+import { type FeedSubmissionFormFormInput } from '../../utils/feed-submission-types';
 
 export const submitNewFeedForm = async (
   formData: FeedSubmissionFormFormInput,
